% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{EZannotations}
\alias{EZannotations}
\title{Download gene annotations}
\usage{
EZannotations()
}
\value{
Downloads the file "PlAnnotation.RDS" to the working directory.
}
\description{
The function downloads the Schmidtea mediterranea gene annotations. The gene symbols consist of Augustus and Blast predictions, respectively.
}
