% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{EZbamSummary}
\alias{EZbamSummary}
\title{Bam File Summary}
\usage{
EZbamSummary()
}
\value{
A csv file named "align_summary.csv" is created in the working directory containing the alignment summary information.
}
\description{
Summarize the alignment statistics for a set BAM files
}
\details{
The function EZBamSummary() reads in summary files with a ".summary"
extension in the working directory and converts them into a neat data frame.
The resulting data frame is then saved as a CSV file with the file name
"align_summary.csv".
}
