% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{EZcol2namerow}
\alias{EZcol2namerow}
\title{Convert Column to Row Names in a Data Frame}
\usage{
EZcol2namerow(df, col = 1)
}
\arguments{
\item{df}{A data frame object.}

\item{col}{An integer specifying which column in the data frame to convert to row names. Default is 1.}
}
\value{
The modified data frame.
}
\description{
Convert a specified column of a data frame into row names, and remove the specified column.
}
\examples{
data(df)
head(df)
EZcol2namerow(df, 1)

}
