% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{EZgene2sym}
\alias{EZgene2sym}
\title{Rename Gene IDs to Gene Symbols}
\usage{
EZgene2sym(df, IDinRow = T)
}
\arguments{
\item{df}{A data frame with gene IDs.}

\item{IDinRow}{A logical indicating whether the gene IDs are in the row names. The default is TRUE.}
}
\value{
A data frame with added gene symbol column, either next to the gene ID column or in the row names.
}
\description{
Map gene IDs to gene symbols in a data frame.
}
\details{
The function takes in a data frame, df, and an optional argument IDinRow. If IDinRow is set to TRUE, the gene IDs are assumed to be in the row names, and the symbol column is added to the data frame. If IDinRow is set to FALSE, the gene IDs are assumed to be in a column, and the symbol column is added next to it.
}
\examples{
df_mapped <- EZgene2sym(df, IDinRow = FALSE)

}
