% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{EZgenePredict}
\alias{EZgenePredict}
\title{Acquire Gene Predictions}
\usage{
EZgenePredict()
}
\value{
This function does not return any values.
}
\description{
Match pre-formatted saf file RNASEQ data (SMESG-high confidence release 1) that complements the genome from EZgenome( ).
}
