% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{EZgenome}
\alias{EZgenome}
\title{Download Genome}
\usage{
EZgenome()
}
\value{
Returns a .fa file containing genome sequence.
}
\description{
Download the planmine Schmidtea mediterranea flatworm genome (SMESG.1) pre-formatted with new lines.
}
