% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{EZhm_order}
\alias{EZhm_order}
\title{Order Heatmap Rows}
\usage{
EZhm_order(df, cols = c(1), decr = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{cols}{A vector of column indices to use for ordering. The default is the first column (1).}

\item{decr}{A logical indicating whether the ordering should be in decreasing order. The default is FALSE.}
}
\value{
A data frame with the same columns but sorted rows based on the mean and standard deviation of the specified columns.
}
\description{
Organize data using z-score.
}
\details{
The 'EZhm_order' function takes a data frame (df) and orders the rows based on
the calculated z-score. The selected
columns can be specified with the 'cols' parameter, which is a vector of
column indices. The default value is the first column (1). The order of
the rows can be either increasing (default) or decreasing, which is specified
with the 'decr' parameter. If multiple columns are selected, the function
calculates the z-score of each row relative to the entire data frame, averages
the z-scores of the selected columns, and sorts the data frame based on these values.
If only one column is selected, the function orders the data frame based on the z-scores
of that single column.
}
