% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{EZindex}
\alias{EZindex}
\title{Generate an Index}
\usage{
EZindex()
}
\description{
Create an index for the reference genome.
}
\details{
If the reference genome is not present in the current working directory, it runs
the EZgenome() function to download it. The index is then built using the
reference genome. This index can be used for further analysis of the
reference genome by aiding in retrieval of desired sequences.

(Note: This step will take 15-20 minutes.)
}
