% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{EZSumTrans}
\alias{EZSumTrans}
\title{Sum Transcript Counts to Gene Counts}
\usage{
EZSumTrans(counts)
}
\arguments{
\item{counts}{A data frame of count data}
}
\value{
A data frame with summed counts for each sample, grouped by 'GeneID', and with row names as column names.
}
\description{
Summarize Transcript Counts by Gene ID
}
\details{
The EZSumTrans function takes a 'counts' data frame and merges it with a saved data frame 'PlAnnotation.RDS' based on a common column. The merged data is then grouped by the 'GeneID' column and the sum of each sample column is calculated. The resulting data is converted to a data frame and the column names are converted to row names
}
